fix(security): 修复 CodeQL 数值乘法溢出告警
在图形命令缓冲、头像缓存、语音 RMS 和表情碰撞数据偏移中先提升乘法运算类型，消除 CodeQL 报告的四个高严重度整数或浮点转换告警。

+++ src/engine/client/graphics_threaded.h
@@ -804,7 +804,7 @@ public:
 
 		// 只延长当前命令缓冲中连续的尾批次，不改变图元顺序。
 		const auto *pPreviousEnd = reinterpret_cast<const unsigned char *>(pPrevious->m_pVertices) +
-					   pPrevious->m_PrimCount * VerticesPerPrimitive * sizeof(SVertex);
+					   static_cast<size_t>(pPrevious->m_PrimCount) * VerticesPerPrimitive * sizeof(SVertex);
 		if(pPreviousEnd != reinterpret_cast<const unsigned char *>(Command.m_pVertices))
 			return false;
 		pPrevious->m_PrimCount += Command.m_PrimCount;
